Moved `Scene` and `Stage` to their own files
